allmost done

+++ Infrastructure/ServiceInterfaces/IExplodable.cs
@@ -1,4 +1,5 @@
 ﻿using Infrastructure.ObjectModel;
+using Microsoft.Xna.Framework;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,7 +10,6 @@ namespace Infrastructure.ServiceInterfaces
 {
     public interface IExplodable
     {
-        int ExplosionRange { get; }
-        //void Explode(int i_XToExplode, int i_YToExplode, Sprite i_SpriteToExplode);
+        float ExplosionRange { get; }
     }
 }
